big group checkin, getting the pipeline back into working shape while using the new registry format.

+++ evaluations/src/makefileEvalWrapper.sh
@@ -1,42 +1,46 @@
 #/bin/bash
+# makefileEvalWrapper.sh
+# 19 Oct 2011
+# dent earl, dearl (a) soe ucsc edu
 #
-#
+# This script is called by the project Makefile when performing
+# an analysis. This script calls a single evaluation on a single
+# prediction and passes along all the proper variables.
+# It also verifies that necessary files exist and fails if they 
+# do not.
 ##########
 set -e
 set -o pipefail
 
+program=$0
+
 evalAndPred=$1
 location=$(readlink -f $2)
+if [ $3 == "registries/.reg.tab" ]; then
+    echo ERROR, you forgot to specify the \'set\' variable. Valid \'set\' options are testSet primates mammals drosophila >&2
+    exit 1
+fi
 registry=$(readlink -f $3)
 tmpDir=$(readlink -f $4)
 outDir=$(readlink -f $5)
 
+if [ -z $location ]; then
+    echo ERROR, specify 'location' >&2
+    exit 1
+fi
+if [ ! -f $registry ]; then
+    echo ERROR, "registry=$registry" not found! Valid \'set\' options are testSet primates mammals drosophila >&2
+    exit 1
+fi
+
 eval=$(readlink -f evaluations/bin/$(dirname $evalAndPred | perl -ple 's/.*\/(.*?)\-.*/$1/')*)
 pred=$location/$(dirname $evalAndPred | perl -ple 's/.*\-(.*)/$1/').maf
 
+##############################
 mkdir -p $(dirname $evalAndPred)
-echo eval $eval 
-echo pred $pred 
-echo location $location 
-echo registry $registry 
-echo tmpDir $tmpDir 
-echo outDir $outDir
-rm -rf $tempDir
 
+# call the evaluation
+$eval $location $pred $registry $tmpDir $outDir
+
+rm -rf $tempDir
 exit 0
-# if [ -z ${set} ]; then 
-#     echo "Error, specify variable 'set'" && exit 1; 
-# fi
-# if [[ ${set} != "testSet" && ${set} != "primates" && ${set} != "mammals" && ${set} != "drosophila" ]]; then 
-#     echo "Error, 'set' must be equal to one of 'testSet' 'primates' 'mammals' 'drosophila'" && exit 1; 
-# fi
-# if [ ! -e registries/${set}.reg.tab ]; then echo "Error, unable to find registry file 'registries/${set}.reg.tab'" && exit 1; fi
-# if [ -z ${location} ]; then echo "Error, specify variable 'location'" && exit 1; fi
-# if [ ! -d ${location} ]; then echo "Error, location=${location} does not appear to be a directory" && exit 1; fi
-# mkdir -p $$(dirname $@) 
-# $(find ${evalBinDir} -name $$(dirname $@ | perl -ple 's/.*\/(.*?)\-.*/$$1/')*) \
-#     ${location}/$$(dirname $@ | perl -ple 's/.*\-(.*)/$$1/').maf \
-#     registries/${set}.reg.tab
-#    $(realpath ${tempDir}) \
-# 	${cwd}/$$(dirname $@)
-# rm -rf $tempDir
